refactor(create-cli): modularize yargs parser
