+++ packages/create-cli/src/index.ts
@@ -1,5 +1,4 @@
 #! /usr/bin/env node
-import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
 import { axeSetupBinding } from '@code-pushup/axe-plugin';
 import { coverageSetupBinding } from '@code-pushup/coverage-plugin';
@@ -8,13 +7,8 @@ import { jsPackagesSetupBinding } from '@code-pushup/js-packages-plugin';
 import { jsDocsSetupBinding } from '@code-pushup/jsdocs-plugin';
 import { lighthouseSetupBinding } from '@code-pushup/lighthouse-plugin';
 import { typescriptSetupBinding } from '@code-pushup/typescript-plugin';
-import { parsePluginSlugs, validatePluginSlugs } from './lib/setup/plugins.js';
-import {
-  CI_PROVIDERS,
-  CONFIG_FILE_FORMATS,
-  type PluginSetupBinding,
-  SETUP_MODES,
-} from './lib/setup/types.js';
+import { yargsCli } from './lib/setup/cli-args.js';
+import type { PluginSetupBinding } from './lib/setup/types.js';
 import { runSetupWizard } from './lib/setup/wizard.js';
 
 const bindings: PluginSetupBinding[] = [
@@ -27,42 +21,6 @@ const bindings: PluginSetupBinding[] = [
   jsDocsSetupBinding,
 ];
 
-const argv = await yargs(hideBin(process.argv))
-  .option('dry-run', {
-    type: 'boolean',
-    default: false,
-    describe: 'Preview changes without writing files',
-  })
-  .option('yes', {
-    alias: 'y',
-    type: 'boolean',
-    default: false,
-    describe: 'Skip prompts and use defaults',
-  })
-  .option('config-format', {
-    type: 'string',
-    choices: CONFIG_FILE_FORMATS,
-    describe: 'Config file format (default: auto-detected from project)',
-  })
-  .option('plugins', {
-    type: 'string',
-    describe: 'Comma-separated plugin slugs to include (e.g. eslint,coverage)',
-    coerce: parsePluginSlugs,
-  })
-  .option('mode', {
-    type: 'string',
-    choices: SETUP_MODES,
-    describe: 'Setup mode (default: auto-detected from project)',
-  })
-  .option('ci', {
-    type: 'string',
-    choices: CI_PROVIDERS,
-    describe: 'CI/CD integration (github, gitlab, or none)',
-  })
-  .check(parsed => {
-    validatePluginSlugs(bindings, parsed.plugins);
-    return true;
-  })
-  .parse();
+const argv = await yargsCli(bindings).parse(hideBin(process.argv));
 
 await runSetupWizard(bindings, argv);
